moved more things to content
