+++ cnxarchive/tests/views_folder/test_content.py
@@ -687,3 +687,97 @@ INSERT INTO trees (nodeid, parent_id, title, childorder, is_collated)
         from ...views_folder.content import get_content
         with self.assertRaises(IdentHashShortId) as cm:
             get_content(self.request)
+
+    def test_content_collection_as_html(self):
+        uuid = 'e79ffde3-7fb4-4af3-9ec8-df648b391597'
+        version = '7.1'
+
+        expected = u"""<html xmlns="http://www.w3.org/1999/xhtml">
+  <body>\
+<ul><li><a href="/contents/e79ffde3-7fb4-4af3-9ec8-df648b391597@7.1.html">College Physics</a>\
+<ul><li><a href="/contents/e79ffde3-7fb4-4af3-9ec8-df648b391597@7.1:209deb1f-1a46-4369-9e0d-18674cf58a3e@7.html">Preface</a></li>\
+<li><a>Introduction: The Nature of Science and Physics</a>\
+<ul><li><a href="/contents/e79ffde3-7fb4-4af3-9ec8-df648b391597@7.1:f3c9ab70-a916-4d8c-9256-42953287b4e9@3.html">Introduction to Science and the Realm of Physics, Physical Quantities, and Units</a></li>\
+<li><a href="/contents/e79ffde3-7fb4-4af3-9ec8-df648b391597@7.1:d395b566-5fe3-4428-bcb2-19016e3aa3ce@4.html">Physics: An Introduction</a></li>\
+<li><a href="/contents/e79ffde3-7fb4-4af3-9ec8-df648b391597@7.1:c8bdbabc-62b1-4a5f-b291-982ab25756d7@6.html">Physical Quantities and Units</a></li>\
+<li><a href="/contents/e79ffde3-7fb4-4af3-9ec8-df648b391597@7.1:5152cea8-829a-4aaf-bcc5-c58a416ecb66@7.html">Accuracy, Precision, and Significant Figures</a></li>\
+<li><a href="/contents/e79ffde3-7fb4-4af3-9ec8-df648b391597@7.1:5838b105-41cd-4c3d-a957-3ac004a48af3@5.html">Approximation</a></li></ul></li>\
+<li><a>Further Applications of Newton's Laws: Friction, Drag, and Elasticity</a>\
+<ul><li><a href="/contents/e79ffde3-7fb4-4af3-9ec8-df648b391597@7.1:24a2ed13-22a6-47d6-97a3-c8aa8d54ac6d@2.html">Introduction: Further Applications of Newton’s Laws</a></li>\
+<li><a href="/contents/e79ffde3-7fb4-4af3-9ec8-df648b391597@7.1:ea271306-f7f2-46ac-b2ec-1d80ff186a59@5.html">Friction</a></li>\
+<li><a href="/contents/e79ffde3-7fb4-4af3-9ec8-df648b391597@7.1:26346a42-84b9-48ad-9f6a-62303c16ad41@6.html">Drag Forces</a></li>\
+<li><a href="/contents/e79ffde3-7fb4-4af3-9ec8-df648b391597@7.1:56f1c5c1-4014-450d-a477-2121e276beca@8.html">Elasticity: Stress and Strain</a></li>\
+</ul></li><li><a href="/contents/e79ffde3-7fb4-4af3-9ec8-df648b391597@7.1:f6024d8a-1868-44c7-ab65-45419ef54881@3.html">Atomic Masses</a></li>\
+<li><a href="/contents/e79ffde3-7fb4-4af3-9ec8-df648b391597@7.1:7250386b-14a7-41a2-b8bf-9e9ab872f0dc@2.html">Selected Radioactive Isotopes</a></li>\
+<li><a href="/contents/e79ffde3-7fb4-4af3-9ec8-df648b391597@7.1:c0a76659-c311-405f-9a99-15c71af39325@5.html">Useful Inførmation</a></li>\
+<li><a href="/contents/e79ffde3-7fb4-4af3-9ec8-df648b391597@7.1:ae3e18de-638d-4738-b804-dc69cd4db3a3@5.html">Glossary of Key Symbols and Notation</a></li></ul></li></ul></body>\n</html>\n"""
+
+        # Build the environment
+        self.request.matchdict = {
+            'ident_hash': '{}@{}'.format(uuid, version),
+            }
+        self.request.matched_route = mock.Mock()
+        self.request.matched_route.name = 'content-html'
+
+        # Call the view
+        from ...views_folder.content import get_content_html
+        resp = get_content_html(self.request)
+
+        # Check that the view returns the expected html
+        p = HTMLParser.HTMLParser()
+        self.assertMultiLineEqual(p.unescape(resp.body), expected)
+
+    def test_content_module_as_html(self):
+        uuid = 'd395b566-5fe3-4428-bcb2-19016e3aa3ce'
+        version = '4'
+
+        # Build the request environment.
+        self.request.matchdict = {'ident_hash': "{}@{}".format(uuid, version)}
+        self.request.matched_route = mock.Mock()
+        self.request.matched_route.name = 'content-html'
+
+        # Call the view.
+        from ...views_folder.content import get_content_html
+
+        # Check that the view returns some html
+        resp_body = get_content_html(self.request).body
+        self.assertTrue(resp_body.startswith('<html'))
+
+    @testing.db_connect
+    def test_content_index_html(self, cursor):
+        uuid = 'ae3e18de-638d-4738-b804-dc69cd4db3a3'
+
+        cursor.execute('ALTER TABLE module_files DISABLE TRIGGER ALL')
+        cursor.execute('DELETE FROM module_files')
+        # Insert a file for version 4
+        cursor.execute('''INSERT INTO files (file, media_type) VALUES
+            (%s, 'text/html') RETURNING fileid''', [memoryview('Version 4')])
+        fileid = cursor.fetchone()[0]
+        cursor.execute('''INSERT INTO module_files
+                       (module_ident, fileid, filename) VALUES
+                       (%s, %s, 'index.cnxml.html')''',
+                       [16, fileid])
+        # Insert a file for version 5
+        cursor.execute('''INSERT INTO files (file, media_type) VALUES
+            (%s, 'text/html') RETURNING fileid''', [memoryview('Version 5')])
+        fileid = cursor.fetchone()[0]
+        cursor.execute('''INSERT INTO module_files
+                       (module_ident, fileid, filename) VALUES
+                       (%s, %s, 'index.cnxml.html')''',
+                       [15, fileid])
+        cursor.connection.commit()
+
+        def get_content(version):
+            # Build the request environment
+            self.request.matchdict = {'ident_hash': '{}@{}'.format(uuid, version)}
+            self.request.matched_route = mock.Mock()
+            self.request.matched_route.name = 'content'
+
+            # Call the view
+            from ...views_folder.content import get_content
+            content = get_content(self.request).json_body
+
+            return content.pop('content')
+
+        self.assertEqual(get_content(4), 'Version 4')
+        self.assertEqual(get_content(5), 'Version 5')
